+++ library/config/dataclasses/sd_peft.py
@@ -5,16 +5,12 @@ from .dataset import DatasetConfig
 from .buckets import BucketsConfig
 from .model import ModelConfig
 from .peft import PeftConfig
-from .saving import SavingConfig
-from .huggingface import HuggingFaceConfig
-from .logging import LoggingConfig
 from .performance import PerformanceConfig
 from .loss import LossConfig
 from .regularization import RegularizationConfig
 from .timestep import TimestepConfig
-from .sampling import SamplingConfig
 from .masked_loss import MaskedLossConfig
-from .metadata import MetadataConfig
+from .output import OutputConfig
 
 
 @dataclass
@@ -25,13 +21,9 @@ class SDPeftConfig:
     buckets: BucketsConfig = field(default_factory=BucketsConfig)
     model: ModelConfig = field(default_factory=ModelConfig)
     peft: PeftConfig = field(default_factory=PeftConfig)
-    saving: SavingConfig = field(default_factory=SavingConfig)
-    huggingface: HuggingFaceConfig = field(default_factory=HuggingFaceConfig)
-    logging: LoggingConfig = field(default_factory=LoggingConfig)
+    output: OutputConfig = field(default_factory=OutputConfig)
     performance: PerformanceConfig = field(default_factory=PerformanceConfig)
     loss: LossConfig = field(default_factory=LossConfig)
     regularization: RegularizationConfig = field(default_factory=RegularizationConfig)
     timestep: TimestepConfig = field(default_factory=TimestepConfig)
-    sampling: SamplingConfig = field(default_factory=SamplingConfig)
     masked_loss: MaskedLossConfig = field(default_factory=MaskedLossConfig)
-    metadata: MetadataConfig = field(default_factory=MetadataConfig)