+++ library/config/dataclasses/sdxl_finetune.py
@@ -4,17 +4,13 @@ from .optimizer import OptimizerConfig
 from .dataset import DatasetConfig
 from .model import ModelConfig
 from .sdxl import SDXLConfig
-from .saving import SavingConfig
-from .huggingface import HuggingFaceConfig
-from .logging import LoggingConfig
 from .performance import PerformanceConfig
 from .loss import LossConfig
 from .regularization import RegularizationConfig
 from .timestep import TimestepConfig
-from .sampling import SamplingConfig
 from .masked_loss import MaskedLossConfig
 from .buckets import BucketsConfig
-from .metadata import MetadataConfig
+from .output import OutputConfig
 
 
 @dataclass
@@ -28,14 +24,10 @@ class SDXLFineTuneConfig:
     dataset: DatasetConfig = field(default_factory=DatasetConfig)
     model: ModelConfig = field(default_factory=ModelConfig)
     sdxl: SDXLConfig = field(default_factory=SDXLConfig)
-    saving: SavingConfig = field(default_factory=SavingConfig)
-    huggingface: HuggingFaceConfig = field(default_factory=HuggingFaceConfig)
-    logging: LoggingConfig = field(default_factory=LoggingConfig)
+    output: OutputConfig = field(default_factory=OutputConfig)
     performance: PerformanceConfig = field(default_factory=PerformanceConfig)
     loss: LossConfig = field(default_factory=LossConfig)
     regularization: RegularizationConfig = field(default_factory=RegularizationConfig)
     timestep: TimestepConfig = field(default_factory=TimestepConfig)
-    sampling: SamplingConfig = field(default_factory=SamplingConfig)
     masked_loss: MaskedLossConfig = field(default_factory=MaskedLossConfig)
     buckets: BucketsConfig = field(default_factory=BucketsConfig)
-    metadata: MetadataConfig = field(default_factory=MetadataConfig)